Make progress toward an interactive command processor and game engine.
There's now a Game class that manages the state of the game, and a
command interpreter that matches on regexes.

Terminal support has been cleaned up and made much nicer. There's now a
Buffer object that manages the screen buffer, which really just wraps
the <pre>'s textContent attribute. Terminal can now manipulate this as
it needs to.